LayoutTests/imported/w3c/web-platform-tests/intersection-observer/text-target-expected.txt



PASS IntersectionObserver observing a br element. 
FAIL First rAF. assert_equals: entries.length expected 1 but got 0
FAIL document.scrollingElement.scrollTop = 300 assert_equals: entries.length expected 2 but got 0
FAIL document.scrollingElement.scrollTop = 100 assert_equals: entries.length expected 3 but got 0
